fixed linking error

+++ src/fractals/julia.cu
@@ -22,7 +22,7 @@
 /// @param cImaginary The imaginary part of the constant c for the Julia set.
 template<typename T>
 __global__ void fractal_rendering(
-        unsigned char* pixels, unsigned long size_of_pixels, unsigned int width, unsigned int height,
+        unsigned char* pixels, size_t size_of_pixels, unsigned int width, unsigned int height,
         T zoom_x, T zoom_y, T x_offset, T y_offset,
         Color* d_palette, int paletteSize, T maxIterations, unsigned int* d_total_iterations,
         T cReal, T cImaginary)
