+++ src/fractals/mandelbrot.cu
@@ -20,7 +20,7 @@
 /// @param d_total_iterations Pointer to a global counter to sum total iterations across all threads/blocks.
 template<typename T>
 __global__ void fractal_rendering(
-        unsigned char* pixels, unsigned long size_of_pixels, unsigned int width, unsigned int height,
+        unsigned char* pixels, size_t size_of_pixels, unsigned int width, unsigned int height,
         T zoom_x, T zoom_y, T x_offset, T y_offset,
         Color* d_palette, int paletteSize, T maxIterations, unsigned int* d_total_iterations)
 {
